contracts-governance: clean up the hardhat config (#2989)
* contracts-governance: clean up the hardhat config

Deletes some unused types and cleans up the hh config file

* contracts-governance: more cleanup
