+++ .changeset/empty-eels-move.md
@@ -1,0 +1,5 @@
+---
+'@eth-optimism/contracts-governance': patch
+---
+
+Clean up the hardhat config
