+++ packages/contracts-governance/scripts/deploy-token.ts
@@ -93,6 +93,10 @@ task('deploy-token', 'Deploy governance token and its mint manager contracts')
         console.log(`Got ${tokenReceipt.contractAddress}`)
         throw new Error(`Fatal error! Mismatch of governance token address`)
       }
+    } else if (nonceTokenDeployer !== 0 && governanceTokenCode === '0x') {
+      throw new Error(
+        `Nonce is ${nonceTokenDeployer} without governance token deployed`
+      )
     } else {
       console.log(
         `GovernanceToken already deployed at ${addresses.governanceToken}, skipping`
